Make document context reusable
